Add password encryption for VMware adapter

+++ oscm-app-vmware/javasrc/org/oscm/app/vmware/persistence/APPDataAccessService.java
@@ -306,8 +306,7 @@ public class APPDataAccessService {
             ResultSet resultSet = stmt.executeQuery();
 
             while (resultSet.next()) {
-                credentials.setUserKey(Long.valueOf(resultSet.getString("key"))
-                        .longValue());
+                credentials.setUserKey(Long.valueOf(resultSet.getString("key")));
                 credentials.setUserId(resultSet.getString("id"));
                 credentials.setPassword(resultSet.getString("password"));
             }
@@ -394,8 +393,7 @@ public class APPDataAccessService {
         return ds;
     }
 
-    protected Context getNamingContext(Properties ctxProperties)
-            throws Exception {
+    protected Context getNamingContext(Properties ctxProperties) throws Exception {
         return new InitialContext(ctxProperties);
     }
 }
